// src/interfaces/hit.ts
export interface IHit {
  author?: string;
  comment_text?: string;
  created_at?: string;
  created_at_i?: number;
  num_comments?: number;
  objectID: string;
  parent_id?: string;
  points?: number;
  story_id?: null;
  story_text?: null;
  story_title?: null;
  story_url?: null;
  title?: string;
  url?: string;
  _highlightResult?: any;
  _tags?: any;
}